created new branch for new layout. Using activities instead of fragments.
