+++ app/src/main/res/menu/navigation.xml
@@ -14,4 +14,5 @@
         android:enabled="true"
         android:title="Profile"/>
 
+
 </menu>
